changing level layout

+++ Disaster/Disaster/Assets/Scripts/DamageBar.cs
@@ -7,13 +7,13 @@ public class DamageBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxAmount(int amount)
+    public void SetMaxAmount(float amount)
     {
         slider.maxValue = amount;
         slider.value = amount;
     }
 
-    public void SetDamage(int damage)
+    public void SetDamage(float damage)
     {
         Debug.Log("Changing amount");
         slider.value = damage;
